Initial OSX enabling work done. Fails on linking stage. Still needs to adress external lib's linking issues. Need to add static lib target due to main function hijacking.

+++ public/include/core/utilities/TarrayAdvanced.h
@@ -145,7 +145,7 @@ assert(ptr);
 // Calculate object index in array
 if ( reinterpret_cast<uint64>(ptr) < reinterpret_cast<uint64>(buffer) )
    return false;
-uint32 index = ( reinterpret_cast<uint64>(ptr) - reinterpret_cast<uint64>(buffer) ) / sizeof(obj);
+uint32 index = static_cast<uint32>( reinterpret_cast<uint64>(ptr) - reinterpret_cast<uint64>(buffer) ) / sizeof(obj);
 if (index >= size)
    return false;
